+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-20T05:21:03.322Z",
+    "2026-05-20T07:16:35.691Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {